PGC phenotype files
Created phenotype files for each MDD symptom for cohorts that have symptom level info for cases and controls. When attempting to remove overlap with UK Biobank, appeared to be a series of duplicated IDs across the samples. Wish to discuss this hence the creation of a branch to suggest any edits before finalising pheno files.